@@ -1056,7 +1056,9 @@
         "redhat.vscode-yaml"
     ],
     "dependencies": {
+        "@kubernetes/client-node": "^0.6.0",
         "@types/lodash": "^4.14.113",
+        "@types/request": "^2.47.1",
         "@types/restify": "^5.0.7",
         "@types/shelljs": "^0.7.8",
         "await-notify": "^1.0.1",
@@ -1069,7 +1071,6 @@
         "fuzzysearch": "^1.0.3",
         "graceful-fs": "^4.1.11",
         "js-yaml": "^3.8.2",
-        "k8s": "^0.4.12",
         "lodash": "^4.17.10",
         "mixin-deep": "^1.3.1",
         "mkdirp": "^0.5.1",
@@ -1100,7 +1101,7 @@
         "gulp-tslint": "^8.1.2",
         "mocha": "^4.1.0",
         "tslint": "^5.9.1",
-        "typescript": "^2.2.32",
+        "typescript": "^2.9.2",
         "vscode": "^1.0.17"
     },
     "repository": {
